+++ src/app/modules/informations/informations.page.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AppVersion } from '@ionic-native/app-version/ngx';
 import { BackNavigationService } from 'src/app/services/back-navigation.service';
 import { AbstractBackNavigationPage } from 'src/app/utils/abstract-back-navigation';
+import * as moment from 'moment';
 
 @Component({
   selector: 'app-informations',
@@ -11,6 +12,7 @@ import { AbstractBackNavigationPage } from 'src/app/utils/abstract-back-navigati
 export class InformationsPage extends AbstractBackNavigationPage implements OnInit {
 
   public version: string;
+  public copyrightTime: string;
 
   constructor(
     private appVersion: AppVersion,
@@ -18,6 +20,7 @@ export class InformationsPage extends AbstractBackNavigationPage implements OnIn
   ) {
     super(backNavigationService, { toHome: false, inRoot: false });
     this.version = '';
+    this.copyrightTime = '';
   }
 
   public async ngOnInit() {
@@ -25,6 +28,11 @@ export class InformationsPage extends AbstractBackNavigationPage implements OnIn
     const versionCode = await this.appVersion.getVersionCode();
     const versionNumber = await this.appVersion.getVersionNumber();
     this.version = `${versionNumber} - build ${versionCode}`;
+
+    // Copyright year
+    const today = moment();
+    const start = moment('31/12/2020');
+    this.copyrightTime = `2020${today.isAfter(start) ? ` - ${today.format('YYYY')}` : ''}`;
   }
 
 }